feature: add new module
